refactor(#78): 로비 준비 상태 변경 유스케이스 서비스 분리
